eclassdb
-----------------------
MySQL Program
-----------------------
eclassdb is a relational database written in mysql with three tables; 'workstudy', 'task', & 'junctiontable', each with four columns. Each table has more than one row insert.

The workstudy column names are id, name, password, and status. 

The task column names are id, room, note, and status.

The junctiontable column names are id, workstudy_id (w_id), task_id (t_id), and stamp. This table associates task id's with a workstudy id.

-----------------------
Java Program
-----------------------
A java program is used to read, modify and write data within the database. The code is to check user input and compare with select data already written within a database table. If the user input matches, a status column within a table is changed from false to true. Otherwise, the program and database do not change.

The deploy server's email feature is used to store received email and insert or UPDATE them into the correct table in appropriate order. If there are no requests in queue, the program and database do not change.

When the workstudy accepts a task, the program will initiate transaction to the task and junctiontable table, ALTERing the old data with new. When the workstudy completes a task, the program will UPDATE the task and junctiontable table appropriately.